vine: release lib tasks (#3824)
* vine: remove lib tasks from retrieved list when using tags

* vine: add comment del lib instances when done

* fix leak of resources in vine_task_copy

* delete lists before copy

* vine: do not copy template unless it is needed.

check_worker_against_task does not modify the target task, thus we can
avoid copying over and over for workers that can't fit the library.

* vine: copy directly to list in vine_task_copy

* fix typo

* format

* add dropped function_slots

* do not use buffers with function calls

They are incompatible with the dask vine executor checkpoints.
Need to rethink the combination of buffer and temporary outputs.

+++ taskvine/src/manager/vine_schedule.c
@@ -132,6 +132,9 @@ int check_worker_have_enough_resources(
 
 int check_worker_against_task(struct vine_manager *q, struct vine_worker_info *w, struct vine_task *t)
 {
+	/* THIS FUNCTION SHOULD NOT MODIFY t IN ANY WAY. */
+	/* Otherwise library templates are modified during the run. */
+
 	/* worker has not reported any resources yet */
 	if (w->resources->tag < 0 || w->resources->workers.total < 1) {
 		return 0;
@@ -175,7 +178,7 @@ int check_worker_against_task(struct vine_manager *q, struct vine_worker_info *w
 
 	// if wall time for worker is specified and there's not enough time for task, then not ok
 	if (w->end_time > 0) {
-		double current_time = timestamp_get() / ONE_SECOND;
+		double current_time = ((double)timestamp_get()) / ONE_SECOND;
 		if (t->resources_requested->end > 0 && w->end_time < t->resources_requested->end) {
 			return 0;
 		}
